Refactor tasks, exceptions and GUI classes into respective folders
All files were in one package and was disorganized, doing this helps with looking for the correct classes/files and groups relevant files together
