+++ src/main/java/Eddie/Exceptions/MissingByDateException.java
@@ -1,9 +1,11 @@
-package Eddie;
+package Eddie.Exceptions;
+
+import Eddie.Exceptions.DukeException;
 
 /**
  * Excpetion when user does not provide by date
  */
-public class MissingByDateException extends DukeException{
+public class MissingByDateException extends DukeException {
     public MissingByDateException() {
         super("Please use /by to include a due date");
     }
