+++ src/main/java/Eddie/GUI/DialogBox.java
@@ -1,4 +1,4 @@
-package Eddie;
+package Eddie.GUI;
 import javafx.geometry.Pos;
 import javafx.scene.control.Label;
 import javafx.scene.image.ImageView;
