+++ src/main/java/Eddie/Tasks/Event.java
@@ -1,4 +1,4 @@
-package Eddie;
+package Eddie.Tasks;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
