+++ src/main/java/Eddie/Tasks/Todo.java
@@ -1,9 +1,11 @@
-package Eddie;
+package Eddie.Tasks;
+
+import Eddie.Tasks.Task;
 
 /**
  * Represents a task to be done no date specified.
  */
-public class Todo extends Task{
+public class Todo extends Task {
 
     public Todo(String name) {
         super(name);
