+++ src/test/java/Eddie/TodoTest.java
@@ -1,4 +1,5 @@
 package Eddie;
+import Eddie.Tasks.Todo;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
